my_automation/requirements.txt
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.2.1
google-genai==1.65.0
google-auth==2.49.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0
google-api-python-client==2.187.0
PyPDF2==3.0.1
pdfplumber==0.11.9
PyMuPDF==1.27.1
gunicorn==21.2.0
whitenoise==6.7.0
pillow==12.1.0
requests==2.32.5
